backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis[hiredis]
aioredis
pydantic
pydantic-settings
email-validator
httpx
jinja2
pandas
openpyxl
